Fix q18_opt SF0.01-incorrect-result & print-plan issues (haog/gqe-python!230)
This MR fixes two issues:

1. the incorrect result issue for `18_opt` SF0.01, this is because we used int instead of double to represent the scale factor in `18_opt`
2. the `q18_opt`'s plan cannot be printed due to missing output data types in `user_defined` relation, this depends on [gqe_MR!432](https://gitlab-master.nvidia.com/Devtech-Compute/gqe/-/merge_requests/432)

How tested:

After the fixes, the `18_opt` plan can be printed without any exception; and enabled the `18_opt` in the CI since it can produce correct results

+++ gqe/benchmark/hardcoded/include/tpch/q18/groupby_task.hpp
@@ -1,6 +1,6 @@
 /*
- * SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
- * SPDX-License-Identifier: LicenseRef-NvidiaProprietary
+ * SPDX-FileCopyrightText: Copyright (c) 2025-2026, NVIDIA CORPORATION & AFFILIATES. All rights
+ * reserved. SPDX-License-Identifier: LicenseRef-NvidiaProprietary
  *
  * NVIDIA CORPORATION, its affiliates and licensors retain all intellectual
  * property and proprietary rights in and to this material, related
@@ -32,7 +32,7 @@ namespace q18 {
  * @return A group by relation with [l_orderkey, sum_l_quantity] where sum_l_quantity > 300
  */
 std::shared_ptr<gqe::physical::relation> groupby(std::shared_ptr<gqe::physical::relation> lineitem,
-                                                 int32_t scale_factor);
+                                                 double scale_factor);
 
 }  // namespace q18
 }  // namespace benchmark
